refactor: create index over utils for simplify references
